# static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "description": "This is the swagger file that goes with our server code",
      "version": "1.0.0",
      "title": "Swagger Rest Api"
    },
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "basePath": "/swagger",
    "paths": {
      "/posts": {
        "get": {
          "operationId": "main.PostsResource.get",
          "tags": [
            "users"
          ],
          "summary": "Read the entire set of users",
          "description": "Read the entire set of users",
          "responses": {
            "200": {
              "description": "Successfully read users set operation",
              "schema": {
                "type": "array",
                "items": {
                  "properties": {
                    "id": {
                      "type": "integer",
                      "description": "Id "
                    },
                    "name": {
                      "type": "string",
                      "description": "First name of the user"
                    },
                    "surname": {
                      "type": "string",
                      "description": "Surname of the user"
                    },
                    "lastname": {
                        "type": "string",
                        "description": "Lastname of the user"
                    },
                      "github_name": {
                        "type": "string",
                        "description": "Github UserName of the user"
                    },
                      "github_follwing": {
                        "type": "string",
                        "description": "Number of people the user is following"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "operationId": "main.PostsResource.post",
          "tags": [
            "users"
          ],
          "summary": "Create an user",
          "description": "Create a new user",
          "requestBody": {
            "required": [
                "name",
                "surname",
                "lastname",
                "github_name"
              ],
            "content": {
                "application/json":{
                    "schema":{
                        "name": {
                            "type": "string"
                            },
                        "surname": {
                            "type": "string"
                            },
                        "lastname": {
                            "type": "string"
                            },
                        "github_name": {
                            "type": "string"
                            }
                        }
                    }
                }
            },
          "responses": {
            "200": {
              "description": "Successfully created user",
              "schema": {
                "properties": {
                  "id": {
                    "type": "integer",
                    "description": "Id of the user"
                  },
                  "name": {
                    "type": "string",
                    "description": "First name of the person"
                  },
                  "surname": {
                    "type": "string",
                    "description": "Surname of the person"
                  },
                  "lastname": {
                      "type": "string",
                      "description": "Lastname of the person"
                  },
                    "github_name": {
                      "type": "string",
                      "description": "Github UserName of the person"
                  },
                    "github_follwing": {
                      "type": "integer",
                      "description": "Number of user the person is following"
                  }
                }
              }
            }
          }
        }
      },
      "/post/{id}": {
        "get": {
          "operationId": "main.PostResource.get",
          "tags": [
            "users"
          ],
          "summary": "Read one user",
          "description": "Read one user",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "Id of the user to get",
              "type": "integer",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "Successfully read user from users data operation",
              "schema": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Id "
                    },
                      "name": {
                        "type": "string",
                        "description": "First name of the person"
                    },
                      "surname": {
                        "type": "string",
                        "description": "Surname of the person"
                    },
                      "lastname": {
                          "type": "string",
                          "description": "Lastname of the person"
                    },
                        "github_name": {
                          "type": "string",
                          "description": "Github UserName of the person"
                    },
                        "github_follwing": {
                          "type": "string",
                          "description": "Number of people the person is following"
                    }
                }
              }
            },
            "400": {
                "description": "Failed. Bad post data."
            }
          }
        },
        "patch": {
          "operationId": "main.PostResource.patch",
          "tags": [
            "users"
          ],
          "summary": "Update a user",
          "description": "Update a user",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "Id the user to update",
              "type": "integer",
              "required": true
            }
          ],
          "requestBody": {
            "required": [
                "name",
                "surname",
                "lastname",
                "github_name"
              ],
            "content": {
                "application/json":{
                    "schema":{
                        "name": {
                            "type": "string"
                            },
                        "surname": {
                            "type": "string"
                            },
                        "lastname": {
                            "type": "string"
                            },
                        "github_name": {
                            "type": "string"
                            }
                        }
                    }
                }
            },
          "responses": {
            "200": {
              "description": "Successfully updated user",
              "schema": {
                "properties": {
                  "id": {
                      "type": "integer",
                      "description": "Id "
                    },
                    "name": {
                        "type": "string",
                        "description": "First name of the person"
                    },
                    "surname": {
                        "type": "string",
                        "description": "Surname of the person"
                    },
                    "lastname": {
                        "type": "string",
                        "description": "Lastname of the person"
                    },
                    "github_name": {
                        "type": "string",
                        "description": "Github UserName of the person"
                    },
                    "github_follwing": {
                        "type": "integer",
                        "description": "Number of people the person is following"
                    }
                }
              }
            }
          }
        },
        "delete": {
          "operationId": "main.PostResource.delete",
          "tags": [
            "users"
          ],
          "summary": "Delete a user from the users list",
          "description": "Delete a user",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "type": "integer",
              "description": "Id of the user to delete",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "Successfully deleted a user"
            }
          }
        }
      }
    }
  }